Workflow changes for trip approval
