+++ app/Models/User.php
@@ -1077,6 +1077,12 @@ class User extends Authenticatable
     {
         $needPayment = $this->assignments()
             ->select('travel_assignments.*')
+            ->whereHas(
+                'travel',
+                static function (Builder $query): void {
+                    $query->whereIn('status', ['approved', 'complete']);
+                }
+            )
             ->unpaid()
             ->oldest('travel.departure_date')
             ->oldest('travel.return_date')
@@ -1088,6 +1094,12 @@ class User extends Authenticatable
 
         $needDocuSign = $this->assignments()
             ->select('travel_assignments.*')
+            ->whereHas(
+                'travel',
+                static function (Builder $query): void {
+                    $query->whereIn('status', ['approved', 'complete']);
+                }
+            )
             ->leftJoin('travel', 'travel.id', '=', 'travel_assignments.travel_id')
             ->needDocuSign()
             ->oldest('travel.departure_date')
@@ -1101,6 +1113,12 @@ class User extends Authenticatable
         // this might be null, but that's fine
         return $this->assignments()
             ->select('travel_assignments.*')
+            ->whereHas(
+                'travel',
+                static function (Builder $query): void {
+                    $query->whereIn('status', ['approved', 'complete']);
+                }
+            )
             ->leftJoin('travel', 'travel.id', '=', 'travel_assignments.travel_id')
             ->oldest('travel.departure_date')
             ->oldest('travel.return_date')
