+++ app/Observers/TravelAssignmentObserver.php
@@ -16,22 +16,26 @@ class TravelAssignmentObserver
 {
     public function created(TravelAssignment $assignment): void
     {
-        if ($assignment->travel->needs_docusign === true) {
-            SendDocuSignEnvelopeForTravelAssignment::dispatch($assignment)
-                ->chain([
-                    new SendTravelAssignmentCreatedNotification($assignment),
-                ]);
-        } else {
-            $assignment->user->notify(new TravelAssignmentCreated($assignment));
-        }
+        if ($assignment->travel->status !== 'draft') {
+            if ($assignment->travel->needs_docusign === true) {
+                SendDocuSignEnvelopeForTravelAssignment::dispatch($assignment, true)
+                    ->chain([
+                        new SendTravelAssignmentCreatedNotification($assignment),
+                    ]);
+            } else {
+                $assignment->user->notify(new TravelAssignmentCreated($assignment));
+            }
 
-        SendTravelAssignmentReminder::dispatch($assignment);
+            SendTravelAssignmentReminder::dispatch($assignment);
+        }
     }
 
     public function saved(TravelAssignment $assignment): void
     {
-        SendTravelAssignmentReminder::dispatch($assignment);
-        PruneTravelAssignmentNotificationsInNova::dispatch($assignment->user);
-        CheckAllTravelAssignmentsComplete::dispatch($assignment->travel);
+        if ($assignment->travel->status !== 'draft') {
+            SendTravelAssignmentReminder::dispatch($assignment);
+            PruneTravelAssignmentNotificationsInNova::dispatch($assignment->user);
+            CheckAllTravelAssignmentsComplete::dispatch($assignment->travel);
+        }
     }
 }
