+++ phpstan.neon
@@ -100,6 +100,7 @@ parameters:
     - '#Method App\\Util\\DocuSign::getAccessToken\(\) should return string but returns mixed\.#'
     - '#Method App\\Util\\Matrix::[a-zA-Z]+\(\) has parameter \$decoded_itinerary with no value type specified in iterable type array\.#'
     - '#Method App\\Util\\Matrix::[a-zA-Z]+\(\) has parameter \$itinerary with no value type specified in iterable type array\.#'
+    - '#Method App\\Util\\SquareCheckout::redirectToSquare\(\) has no return type specified\.#'
     - '#Method class@anonymous\/database\/migrations\/[0-9-a-z_]+.php:[0-9]+::getConnection\(\) should return string but returns mixed\.#'
     - '#Method Database\\Factories\\[a-zA-Z]+::definition\(\) should return array<string, .+> but returns array<string, mixed>\.#'
     - '#Offset ''[A-Za-z]+'' does not exist on array\|null\.#'
@@ -230,7 +231,7 @@ parameters:
     - '#Parameter \#2 \$model_class of job class App\\Jobs\\PushToJedi constructor expects string in App\\Jobs\\PushToJedi::dispatch\(\), class-string<App\\Models\\User>\|null given\.#'
     - '#Parameter \#2 \$name of method Illuminate\\Mail\\Mailable::replyTo\(\) expects string\|null, mixed given\.#'
     - '#Parameter \#2 \$operator of method Illuminate\\Database\\Eloquent\\Builder<App\\Models\\Attendance>::whereDate\(\) expects DateTimeInterface\|string\|null, mixed given\.#'
-    - '#Parameter \#2 \$payment of static method App\\Http\\Controllers\\SquareCheckoutController::redirect\(\) expects App\\Models\\Payment, Illuminate\\Database\\Eloquent\\Model given\.#'
+    - '#Parameter \#2 \$payment of static method App\\Util\\SquareCheckout::redirectToSquare\(\) expects App\\Models\\Payment, Illuminate\\Database\\Eloquent\\Model given\.#'
     - '#Parameter \#2 \$query \(Illuminate\\Database\\Eloquent\\Builder<App\\Models\\[a-zA-Z]+>\) of method App\\Nova\\[a-zA-Z]+::indexQuery\(\) should be contravariant with parameter \$query \(Illuminate\\Database\\Eloquent\\Builder\) of method Laravel\\Nova\\Resource<Illuminate\\Database\\Eloquent\\Model>::indexQuery\(\)#'
     - '#Parameter \#2 \$query \(Illuminate\\Database\\Eloquent\\Builder<App\\Models\\[a-zA-Z]+>\) of method App\\Nova\\[a-zA-Z]+::relatableQuery\(\) should be contravariant with parameter \$query \(Illuminate\\Database\\Eloquent\\Builder\) of method Laravel\\Nova\\Resource<Illuminate\\Database\\Eloquent\\Model>::relatableQuery\(\)#'
     - '#Parameter \#2 \$query \(Illuminate\\Database\\Eloquent\\Builder<App\\Models\\[a-zA-Z]+>\) of method App\\Nova\\Filters\\[a-zA-Z]+::apply\(\) should be contravariant with parameter \$query \(Illuminate\\Database\\Eloquent\\Builder\) of method Laravel\\Nova\\Contracts\\Filter::apply\(\)#'
@@ -245,7 +246,7 @@ parameters:
     - '#Parameter \#3 \$fail of method App\\Rules\\MatrixItineraryBusinessPolicy::validate\(\) expects Closure\(string\): Illuminate\\Translation\\PotentiallyTranslatedString, Closure\(string\): void given\.#'
     - '#Parameter \#3 \$resolveCallback of static method Laravel\\Nova\\Fields\\Field::make\(\) expects \(callable\(\): mixed\)\|null, ''App\\\\Nova\\\\Airport'' given\.#'
     - '#Parameter \#3 \$subject of function str_replace expects array\|string, mixed given\.#'
-    - '#Parameter \#3 \$user of static method App\\Http\\Controllers\\SquareCheckoutController::redirect\(\) expects App\\Models\\User, App\\Models\\User\|null given\.#'
+    - '#Parameter \#3 \$user of static method App\\Util\\SquareCheckout::redirectToSquare\(\) expects App\\Models\\User, App\\Models\\User\|null given\.#'
     - '#Parameter \#3 \$value \(.+\) of method App\\Nova\\Filters\\[a-zA-Z]+::apply\(\) should be contravariant with parameter \$value \(mixed\) of method Laravel\\Nova\\Contracts\\Filter::apply\(\)#'
     - '#Parameter \#3 \$value \(.+\) of method App\\Nova\\Filters\\[a-zA-Z]+::apply\(\) should be contravariant with parameter \$value \(mixed\) of method Laravel\\Nova\\Filters\\Filter::apply\(\)#'
     - '#Parameter \#5 \$column of method Laravel\\Nova\\Metrics\\Trend::aggregate\(\) expects Illuminate\\Database\\Query\\Expression\|string, Illuminate\\Contracts\\Database\\Query\\Expression given\.#'
