+++ psalm.xml
@@ -61,12 +61,9 @@
         <MissingReturnType>
             <errorLevel type="suppress">
                 <directory name="app/Http/Controllers" />
-            </errorLevel>
-            <errorLevel type="suppress">
                 <directory name="app/Http/Middleware" />
-            </errorLevel>
-            <errorLevel type="suppress">
                 <directory name="app/Nova/Actions" />
+                <file name="app/Util/SquareCheckout.php" />
             </errorLevel>
         </MissingReturnType>
         <UnresolvableInclude>
@@ -169,6 +166,7 @@
         <PossiblyUndefinedVariable>
             <errorLevel type="suppress">
                 <file name="app/Http/Controllers/SquareCheckoutController.php" />
+                <file name="app/Util/SquareCheckout.php" />
                 <file name="app/Nova/User.php" />
             </errorLevel>
         </PossiblyUndefinedVariable>
